+++ components/cards/ShinjukuVisitorsCard.vue
@@ -1,6 +1,6 @@
 <template>
   <v-col cols="12" md="6" class="DataCard">
-    <shinjuku-bar-chart
+    <visitors-bar-chart
       :title="$t('新宿区エリアの来訪者推移（参考値）')"
       :title-id="'shinjuku-visitors'"
       :chart-id="'shinjuku-visitors'"
@@ -15,15 +15,15 @@
 
 <script>
 import ShinjukuData from '@/data/13104_daily_visitors.json'
-import ShinjukuBarChart from '@/components/ShinjukuBarChart.vue'
+import VisitorsBarChart from '@/components/VisitorsBarChart.vue'
 
 export default {
   components: {
-    ShinjukuBarChart
+    VisitorsBarChart
   },
   data() {
     // ツールチップ label 文字列
-    // this.$t を使うため shinjuku-bar-chart の外側へ
+    // this.$t を使うため visitors-bar-chart の外側へ
     const tooltipTitle = tooltipItems => {
       const label = tooltipItems[0].label
       return this.$t('期間: {duration}', {
