+++ pages/cards/_card.vue
@@ -36,7 +36,9 @@
       "
     />
     <agency-card v-else-if="this.$route.params.card == 'agency'" />
-    <shinjuku-card v-else-if="this.$route.params.card == 'shinjuku-visitors'" />
+    <shinjuku-visitors-card
+      v-else-if="this.$route.params.card == 'shinjuku-visitors'"
+    />
   </div>
 </template>
 
@@ -55,7 +57,7 @@ import TelephoneAdvisoryReportsNumberCard from '@/components/cards/TelephoneAdvi
 import ConsultationDeskReportsNumberCard from '@/components/cards/ConsultationDeskReportsNumberCard.vue'
 import MetroCard from '@/components/cards/MetroCard.vue'
 import AgencyCard from '@/components/cards/AgencyCard.vue'
-import ShinjukuCard from '@/components/cards/ShinjukuCard.vue'
+import ShinjukuVisitorsCard from '@/components/cards/ShinjukuVisitorsCard.vue'
 
 export default {
   components: {
@@ -69,7 +71,7 @@ export default {
     ConsultationDeskReportsNumberCard,
     MetroCard,
     AgencyCard,
-    ShinjukuCard
+    ShinjukuVisitorsCard
   },
   data() {
     let title, updatedAt
