+++ pages/index.vue
@@ -23,7 +23,7 @@
       <consultation-desk-reports-number-card />
       <metro-card />
       <agency-card />
-      <shinjuku-card />
+      <shinjuku-visitors-card />
     </v-row>
   </div>
 </template>
@@ -46,7 +46,7 @@ import TelephoneAdvisoryReportsNumberCard from '@/components/cards/TelephoneAdvi
 import ConsultationDeskReportsNumberCard from '@/components/cards/ConsultationDeskReportsNumberCard.vue'
 import MetroCard from '@/components/cards/MetroCard.vue'
 import AgencyCard from '@/components/cards/AgencyCard.vue'
-import ShinjukuCard from '@/components/cards/ShinjukuCard.vue'
+import ShinjukuVisitorsCard from '@/components/cards/ShinjukuVisitorsCard.vue'
 
 export default Vue.extend({
   components: {
@@ -63,7 +63,7 @@ export default Vue.extend({
     ConsultationDeskReportsNumberCard,
     MetroCard,
     AgencyCard,
-    ShinjukuCard
+    ShinjukuVisitorsCard
   },
   data() {
     const data = {
